Refactor code and move tests to dedicated modules
Moved certain function definitions to new dedicated modules such as 'packages.rs' and 'packed_crate.rs'. Unit tests previously embedded within sources are moved to designated testing files. Code is refactored to accommodate these changes. This refactoring improves code organization, readability, and maintainability.
