+++ module/move/willbe/src/endpoint/list.rs
@@ -11,11 +11,7 @@ mod private
   };
   use std::str::FromStr;
   // use anyhow::Context;
-  use package::
-  {
-    FilterMapOptions,
-    packages_filter_map
-  };
+  use packages::FilterMapOptions;
   use wtools::error::
   {
     for_app::{ Error, Context },
@@ -107,9 +103,18 @@ mod private
   }
 
   /// Wrapper to redirect output from `ptree` graph to `fmt::Write`
-  pub( crate ) struct Io2FmtWrite< 'a, W >
+  pub struct Io2FmtWrite< 'a, W >
   {
+    /// This struct provides a mutable reference to a writer and is used as a formatting writer.
     pub f : &'a mut W,
+  }
+
+  impl< T > std::fmt::Debug for Io2FmtWrite< '_, T >
+  {
+    fn fmt( &self, f : &mut Formatter< '_ > ) -> std::fmt::Result
+    {
+      f.debug_struct( std::any::type_name::< Self >() ).finish()
+    }
   }
 
   impl< W : std::fmt::Write > std::io::Write for Io2FmtWrite< '_, W >
@@ -194,7 +199,7 @@ mod private
       }
     };
 
-    let packages_map =  packages_filter_map
+    let packages_map =  packages::filter
     (
       &metadata.load().packages_get(),
       FilterMapOptions{ dependency_filter: dep_filter, ..Default::default() }
@@ -272,32 +277,6 @@ mod private
   }
 }
 
-// qqq : for Bohdan : move to tests folder
-mod tests
-{
-  #[ test ]
-  fn io2fmt_write()
-  {
-    use super::private::Io2FmtWrite;
-
-    // Arrange
-    fn accepts_io_write< W : std::io::Write >( mut w : W ) -> std::io::Result< () >
-    {
-      w.write( b"Hello, world!" )?;
-
-      Ok( () )
-    }
-
-    let mut string = String::new();
-
-    // Act
-    accepts_io_write( Io2FmtWrite { f : &mut string } ).unwrap();
-
-    // Assert
-    assert_eq!( "Hello, world!", &string );
-  }
-}
-
 //
 
 crate::mod_interface!
@@ -310,4 +289,6 @@ crate::mod_interface!
   protected use ListReport;
   /// List packages in workspace.
   orphan use list;
+  /// Wrapper to redirect output from `io::Write` to `fmt::Write`
+  protected use Io2FmtWrite;
 }
