+++ module/move/willbe/tests/inc/mod.rs
@@ -4,3 +4,5 @@ mod dependencies;
 mod endpoints;
 mod publish_need;
 mod query;
+mod version;
+mod io2fmt_write;
